refactor(mcp): extract diff, step, validation into tools/ modules
- tools/diff.rs: diff_libraries + diff_pcblibs/schlibs
- tools/step.rs: extract_step_model + list/extract_all/by_footprint/output
  (moves the base64 import here too)
- tools/validation.rs: coordinate/footprint/symbol validators

server.rs: ~9890 -> 8924 lines. Build/clippy/tests green.
