+++ src/mcp/server.rs
@@ -13,7 +13,6 @@
 
 use std::path::PathBuf;
 
-use base64::{engine::general_purpose::STANDARD as BASE64_STANDARD, Engine as _};
 use serde::{Deserialize, Serialize};
 use serde_json::{json, Value};
 
@@ -2024,214 +2023,6 @@ impl McpServer {
             .map_or(0.0, |(key, _)| key as f64 / 100.0)
     }
 
-    // ==================== Coordinate Validation ====================
-
-    /// Maximum coordinate value in mm that can be safely converted to Altium internal units.
-    /// Internal units use i32: max value ~5456 mm (`i32::MAX` / 393700.7874).
-    /// We use 5000 mm (~200 inches) as a conservative limit.
-    const MAX_COORDINATE_MM: f64 = 5000.0;
-
-    /// Validates that a coordinate is within the safe range for Altium internal units.
-    fn validate_coordinate(value: f64, field_name: &str) -> Result<(), String> {
-        if !value.is_finite() {
-            return Err(format!(
-                "{field_name} must be a finite number, got: {value}"
-            ));
-        }
-        if value.abs() > Self::MAX_COORDINATE_MM {
-            return Err(format!(
-                "{field_name} value {value} mm exceeds the maximum safe range of ±{} mm",
-                Self::MAX_COORDINATE_MM
-            ));
-        }
-        Ok(())
-    }
-
-    /// Checks if a pad's per-layer data is uniform (all layers have same values).
-    ///
-    /// When all per-layer values are identical, the data is redundant and can be
-    /// omitted in compact mode, even if the pad was stored with `FullStack` mode.
-    fn pad_has_uniform_per_layer_data(pad: &crate::altium::pcblib::Pad) -> bool {
-        // Check per_layer_sizes - all should match primary width/height
-        let sizes_uniform = pad.per_layer_sizes.as_ref().map_or(true, |sizes| {
-            sizes
-                .iter()
-                .all(|&(w, h)| (w - pad.width).abs() < 0.001 && (h - pad.height).abs() < 0.001)
-        });
-
-        // Check per_layer_shapes - all should match primary shape
-        let shapes_uniform = pad
-            .per_layer_shapes
-            .as_ref()
-            .map_or(true, |shapes| shapes.iter().all(|s| *s == pad.shape));
-
-        // Check per_layer_corner_radii - all should match primary corner_radius_percent
-        let primary_radius = pad.corner_radius_percent.unwrap_or(0);
-        let radii_uniform = pad
-            .per_layer_corner_radii
-            .as_ref()
-            .map_or(true, |radii| radii.iter().all(|&r| r == primary_radius));
-
-        // Check per_layer_offsets - all should be zero (no offset)
-        let offsets_uniform = pad.per_layer_offsets.as_ref().map_or(true, |offsets| {
-            offsets
-                .iter()
-                .all(|&(x, y)| x.abs() < 0.001 && y.abs() < 0.001)
-        });
-
-        sizes_uniform && shapes_uniform && radii_uniform && offsets_uniform
-    }
-
-    /// Validates all coordinates in a footprint before writing.
-    fn validate_footprint_coordinates(
-        footprint: &crate::altium::pcblib::Footprint,
-    ) -> Result<(), String> {
-        let name = &footprint.name;
-
-        for (i, pad) in footprint.pads.iter().enumerate() {
-            Self::validate_coordinate(pad.x, &format!("Footprint '{name}' pad {i} x"))?;
-            Self::validate_coordinate(pad.y, &format!("Footprint '{name}' pad {i} y"))?;
-            Self::validate_coordinate(pad.width, &format!("Footprint '{name}' pad {i} width"))?;
-            Self::validate_coordinate(pad.height, &format!("Footprint '{name}' pad {i} height"))?;
-            if let Some(hole) = pad.hole_size {
-                Self::validate_coordinate(hole, &format!("Footprint '{name}' pad {i} hole_size"))?;
-            }
-        }
-
-        for (i, track) in footprint.tracks.iter().enumerate() {
-            Self::validate_coordinate(track.x1, &format!("Footprint '{name}' track {i} x1"))?;
-            Self::validate_coordinate(track.y1, &format!("Footprint '{name}' track {i} y1"))?;
-            Self::validate_coordinate(track.x2, &format!("Footprint '{name}' track {i} x2"))?;
-            Self::validate_coordinate(track.y2, &format!("Footprint '{name}' track {i} y2"))?;
-            Self::validate_coordinate(track.width, &format!("Footprint '{name}' track {i} width"))?;
-        }
-
-        for (i, arc) in footprint.arcs.iter().enumerate() {
-            Self::validate_coordinate(arc.x, &format!("Footprint '{name}' arc {i} x"))?;
-            Self::validate_coordinate(arc.y, &format!("Footprint '{name}' arc {i} y"))?;
-            Self::validate_coordinate(arc.radius, &format!("Footprint '{name}' arc {i} radius"))?;
-            Self::validate_coordinate(arc.width, &format!("Footprint '{name}' arc {i} width"))?;
-        }
-
-        for (i, region) in footprint.regions.iter().enumerate() {
-            for (j, vertex) in region.vertices.iter().enumerate() {
-                Self::validate_coordinate(
-                    vertex.x,
-                    &format!("Footprint '{name}' region {i} vertex {j} x"),
-                )?;
-                Self::validate_coordinate(
-                    vertex.y,
-                    &format!("Footprint '{name}' region {i} vertex {j} y"),
-                )?;
-            }
-        }
-
-        for (i, text) in footprint.text.iter().enumerate() {
-            Self::validate_coordinate(text.x, &format!("Footprint '{name}' text {i} x"))?;
-            Self::validate_coordinate(text.y, &format!("Footprint '{name}' text {i} y"))?;
-            Self::validate_coordinate(text.height, &format!("Footprint '{name}' text {i} height"))?;
-        }
-
-        Ok(())
-    }
-
-    /// Maximum coordinate value for `SchLib` (uses i16 internally).
-    /// `i16::MAX` = 32767, but we use 32000 as a conservative limit.
-    const MAX_SCHLIB_COORDINATE: i32 = 32000;
-
-    /// Validates that a `SchLib` coordinate is within the safe range for i16.
-    fn validate_schlib_coordinate(value: i32, field_name: &str) -> Result<(), String> {
-        if value.abs() > Self::MAX_SCHLIB_COORDINATE {
-            return Err(format!(
-                "{field_name} value {value} exceeds the maximum safe range of ±{} units",
-                Self::MAX_SCHLIB_COORDINATE
-            ));
-        }
-        Ok(())
-    }
-
-    /// Validates all coordinates in a symbol before writing.
-    fn validate_symbol_coordinates(symbol: &crate::altium::schlib::Symbol) -> Result<(), String> {
-        let name = &symbol.name;
-
-        for (i, pin) in symbol.pins.iter().enumerate() {
-            Self::validate_schlib_coordinate(pin.x, &format!("Symbol '{name}' pin {i} x"))?;
-            Self::validate_schlib_coordinate(pin.y, &format!("Symbol '{name}' pin {i} y"))?;
-            Self::validate_schlib_coordinate(
-                pin.length,
-                &format!("Symbol '{name}' pin {i} length"),
-            )?;
-        }
-
-        for (i, rect) in symbol.rectangles.iter().enumerate() {
-            Self::validate_schlib_coordinate(
-                rect.x1,
-                &format!("Symbol '{name}' rectangle {i} x1"),
-            )?;
-            Self::validate_schlib_coordinate(
-                rect.y1,
-                &format!("Symbol '{name}' rectangle {i} y1"),
-            )?;
-            Self::validate_schlib_coordinate(
-                rect.x2,
-                &format!("Symbol '{name}' rectangle {i} x2"),
-            )?;
-            Self::validate_schlib_coordinate(
-                rect.y2,
-                &format!("Symbol '{name}' rectangle {i} y2"),
-            )?;
-        }
-
-        for (i, line) in symbol.lines.iter().enumerate() {
-            Self::validate_schlib_coordinate(line.x1, &format!("Symbol '{name}' line {i} x1"))?;
-            Self::validate_schlib_coordinate(line.y1, &format!("Symbol '{name}' line {i} y1"))?;
-            Self::validate_schlib_coordinate(line.x2, &format!("Symbol '{name}' line {i} x2"))?;
-            Self::validate_schlib_coordinate(line.y2, &format!("Symbol '{name}' line {i} y2"))?;
-        }
-
-        for (i, polyline) in symbol.polylines.iter().enumerate() {
-            for (j, &(x, y)) in polyline.points.iter().enumerate() {
-                Self::validate_schlib_coordinate(
-                    x,
-                    &format!("Symbol '{name}' polyline {i} point {j} x"),
-                )?;
-                Self::validate_schlib_coordinate(
-                    y,
-                    &format!("Symbol '{name}' polyline {i} point {j} y"),
-                )?;
-            }
-        }
-
-        for (i, arc) in symbol.arcs.iter().enumerate() {
-            Self::validate_schlib_coordinate(arc.x, &format!("Symbol '{name}' arc {i} x"))?;
-            Self::validate_schlib_coordinate(arc.y, &format!("Symbol '{name}' arc {i} y"))?;
-            Self::validate_schlib_coordinate(
-                arc.radius,
-                &format!("Symbol '{name}' arc {i} radius"),
-            )?;
-        }
-
-        for (i, ellipse) in symbol.ellipses.iter().enumerate() {
-            Self::validate_schlib_coordinate(ellipse.x, &format!("Symbol '{name}' ellipse {i} x"))?;
-            Self::validate_schlib_coordinate(ellipse.y, &format!("Symbol '{name}' ellipse {i} y"))?;
-            Self::validate_schlib_coordinate(
-                ellipse.radius_x,
-                &format!("Symbol '{name}' ellipse {i} radius_x"),
-            )?;
-            Self::validate_schlib_coordinate(
-                ellipse.radius_y,
-                &format!("Symbol '{name}' ellipse {i} radius_y"),
-            )?;
-        }
-
-        for (i, label) in symbol.labels.iter().enumerate() {
-            Self::validate_schlib_coordinate(label.x, &format!("Symbol '{name}' label {i} x"))?;
-            Self::validate_schlib_coordinate(label.y, &format!("Symbol '{name}' label {i} y"))?;
-        }
-
-        Ok(())
-    }
-
     // ==================== Library Management Tools ====================
 
     /// Deletes one or more components from a library file.
@@ -3540,765 +3331,6 @@ impl McpServer {
         if let Some(validation) = Self::post_write_validation_schlib(output_path) {
             result["validation"] = validation;
         }
-
-        ToolCallResult::text(serde_json::to_string_pretty(&result).unwrap())
-    }
-
-    // ==================== STEP Model Extraction ====================
-
-    /// Extracts embedded STEP 3D models from a `PcbLib` file.
-    #[allow(clippy::too_many_lines, clippy::cast_possible_truncation)]
-    fn call_extract_step_model(&self, arguments: &Value) -> ToolCallResult {
-        use crate::altium::PcbLib;
-
-        let Some(filepath) = arguments.get("filepath").and_then(Value::as_str) else {
-            return ToolCallResult::error("Missing required parameter: filepath");
-        };
-
-        // Validate library path
-        if let Err(e) = self.validate_path(filepath) {
-            return ToolCallResult::error(e);
-        }
-
-        // Get extraction mode (default: auto)
-        let mode = arguments
-            .get("mode")
-            .and_then(Value::as_str)
-            .unwrap_or("auto");
-
-        // Validate output path if provided
-        let output_path = arguments.get("output_path").and_then(Value::as_str);
-        if let Some(out_path) = output_path {
-            if let Err(e) = self.validate_path(out_path) {
-                return ToolCallResult::error(e);
-            }
-        }
-
-        let model_identifier = arguments.get("model").and_then(Value::as_str);
-        let footprint_name = arguments.get("footprint_name").and_then(Value::as_str);
-
-        // Parse optional pagination parameters (for listing models)
-        let limit = arguments
-            .get("limit")
-            .and_then(Value::as_u64)
-            .map(|v| v as usize);
-        let offset = arguments
-            .get("offset")
-            .and_then(Value::as_u64)
-            .map_or(0, |v| v as usize);
-
-        // Read the library
-        let library = match PcbLib::open(filepath) {
-            Ok(lib) => lib,
-            Err(e) => {
-                let result = json!({
-                    "status": "error",
-                    "filepath": filepath,
-                    "error": e.to_string(),
-                });
-                return ToolCallResult::error(serde_json::to_string_pretty(&result).unwrap());
-            }
-        };
-
-        // Get all embedded models
-        let models: Vec<_> = library.models().collect();
-
-        if models.is_empty() {
-            // Check for external model references to provide better error context
-            let external_refs: Vec<Value> = library
-                .iter()
-                .filter_map(|fp| {
-                    fp.model_3d.as_ref().map(|m| {
-                        json!({
-                            "footprint": fp.name,
-                            "filepath": m.filepath,
-                        })
-                    })
-                })
-                .collect();
-
-            let component_body_refs: Vec<Value> = library
-                .iter()
-                .filter(|fp| !fp.component_bodies.is_empty())
-                .map(|fp| {
-                    json!({
-                        "footprint": fp.name,
-                        "body_count": fp.component_bodies.len(),
-                        "model_ids": fp.component_bodies.iter().map(|cb| &cb.model_id).collect::<Vec<_>>(),
-                    })
-                })
-                .collect();
-
-            let mut result = json!({
-                "status": "error",
-                "filepath": filepath,
-                "error": "No embedded 3D models found in this library.",
-            });
-
-            if !external_refs.is_empty() {
-                result["note"] = json!("This library uses external STEP file references (not embedded). The model files are stored separately on disk.");
-                result["external_model_references"] = json!(external_refs);
-            }
-
-            if !component_body_refs.is_empty() {
-                result["component_body_references"] = json!(component_body_refs);
-                result["note"] = json!("Component bodies reference model IDs, but the corresponding model data was not found in /Library/Models/. The models may have been removed or the library may be using external references.");
-            }
-
-            if external_refs.is_empty() && component_body_refs.is_empty() {
-                result["note"] = json!("No 3D model references of any kind found in this library.");
-            }
-
-            return ToolCallResult::error(serde_json::to_string_pretty(&result).unwrap());
-        }
-
-        // Handle different modes
-        match mode {
-            "list" => {
-                // Force list mode - always list models
-                Self::list_step_models(filepath, &models, limit, offset)
-            }
-            "extract_all" => {
-                // Extract all models to output directory
-                Self::extract_all_step_models(filepath, output_path, &models)
-            }
-            "extract_by_footprint" => {
-                // Extract models used by a specific footprint
-                let Some(fp_name) = footprint_name else {
-                    return ToolCallResult::error(
-                        "Missing required parameter 'footprint_name' for extract_by_footprint mode",
-                    );
-                };
-                Self::extract_step_by_footprint(filepath, output_path, &library, &models, fp_name)
-            }
-            _ => {
-                // Default "auto" mode - original behaviour
-                let target_model = if let Some(identifier) = model_identifier {
-                    // Look up by name or GUID
-                    models
-                        .iter()
-                        .find(|m| {
-                            m.name.eq_ignore_ascii_case(identifier)
-                                || m.id.eq_ignore_ascii_case(identifier)
-                                || m.id
-                                    .trim_matches(|c| c == '{' || c == '}')
-                                    .eq_ignore_ascii_case(identifier)
-                        })
-                        .copied()
-                } else if models.len() == 1 {
-                    // Only one model, extract it
-                    Some(models[0])
-                } else {
-                    // Multiple models, list them with pagination
-                    return Self::list_step_models(filepath, &models, limit, offset);
-                };
-
-                let Some(model) = target_model else {
-                    let model_list: Vec<Value> = models
-                        .iter()
-                        .map(|m| {
-                            json!({
-                                "id": m.id,
-                                "name": m.name,
-                            })
-                        })
-                        .collect();
-
-                    let result = json!({
-                        "status": "error",
-                        "filepath": filepath,
-                        "error": format!("Model '{}' not found.", model_identifier.unwrap_or("")),
-                        "available_models": model_list,
-                    });
-                    return ToolCallResult::error(serde_json::to_string_pretty(&result).unwrap());
-                };
-
-                // Extract the model - write to file or return as base64
-                Self::extract_model_output(filepath, output_path, model)
-            }
-        }
-    }
-
-    /// Lists embedded STEP models with pagination.
-    fn list_step_models(
-        filepath: &str,
-        models: &[&crate::altium::pcblib::EmbeddedModel],
-        limit: Option<usize>,
-        offset: usize,
-    ) -> ToolCallResult {
-        let total_count = models.len();
-        let model_list: Vec<Value> = models
-            .iter()
-            .skip(offset)
-            .take(limit.unwrap_or(usize::MAX))
-            .map(|m| {
-                json!({
-                    "id": m.id,
-                    "name": m.name,
-                    "size_bytes": m.data.len(),
-                })
-            })
-            .collect();
-
-        let returned_count = model_list.len();
-        let has_more = offset + returned_count < total_count;
-
-        let result = json!({
-            "status": "list",
-            "filepath": filepath,
-            "total_count": total_count,
-            "returned_count": returned_count,
-            "offset": offset,
-            "has_more": has_more,
-            "models": model_list,
-        });
-        ToolCallResult::text(serde_json::to_string_pretty(&result).unwrap())
-    }
-
-    /// Extracts all STEP models to an output directory.
-    fn extract_all_step_models(
-        filepath: &str,
-        output_path: Option<&str>,
-        models: &[&crate::altium::pcblib::EmbeddedModel],
-    ) -> ToolCallResult {
-        let Some(output_dir) = output_path else {
-            return ToolCallResult::error(
-                "Missing required parameter 'output_path' (directory) for extract_all mode",
-            );
-        };
-
-        // Create output directory if it doesn't exist
-        let out_dir = std::path::Path::new(output_dir);
-        if let Err(e) = std::fs::create_dir_all(out_dir) {
-            return ToolCallResult::error(format!("Failed to create output directory: {e}"));
-        }
-
-        let mut extracted: Vec<Value> = Vec::new();
-        let mut errors: Vec<Value> = Vec::new();
-
-        for model in models {
-            let output_file = out_dir.join(&model.name);
-            match std::fs::write(&output_file, &model.data) {
-                Ok(()) => {
-                    extracted.push(json!({
-                        "id": model.id,
-                        "name": model.name,
-                        "output_path": output_file.to_string_lossy(),
-                        "size_bytes": model.data.len(),
-                    }));
-                }
-                Err(e) => {
-                    errors.push(json!({
-                        "id": model.id,
-                        "name": model.name,
-                        "error": e.to_string(),
-                    }));
-                }
-            }
-        }
-
-        let result = json!({
-            "status": if errors.is_empty() { "success" } else { "partial" },
-            "filepath": filepath,
-            "output_directory": output_dir,
-            "total_models": models.len(),
-            "extracted_count": extracted.len(),
-            "error_count": errors.len(),
-            "extracted": extracted,
-            "errors": errors,
-        });
-        ToolCallResult::text(serde_json::to_string_pretty(&result).unwrap())
-    }
-
-    /// Extracts STEP models used by a specific footprint.
-    fn extract_step_by_footprint(
-        filepath: &str,
-        output_path: Option<&str>,
-        library: &crate::altium::PcbLib,
-        models: &[&crate::altium::pcblib::EmbeddedModel],
-        footprint_name: &str,
-    ) -> ToolCallResult {
-        // Find the footprint
-        let Some(footprint) = library.get(footprint_name) else {
-            let available: Vec<&str> = library.iter().map(|fp| fp.name.as_str()).collect();
-            let result = json!({
-                "status": "error",
-                "filepath": filepath,
-                "error": format!("Footprint '{}' not found", footprint_name),
-                "available_footprints": available,
-            });
-            return ToolCallResult::error(serde_json::to_string_pretty(&result).unwrap());
-        };
-
-        // Collect model IDs used by this footprint
-        let mut model_ids: std::collections::HashSet<String> = std::collections::HashSet::new();
-
-        // From component bodies
-        for cb in &footprint.component_bodies {
-            model_ids.insert(cb.model_id.to_lowercase());
-        }
-
-        // From model_3d reference (if it points to an embedded model)
-        if let Some(ref m3d) = footprint.model_3d {
-            if !m3d.filepath.is_empty() {
-                // Try to match by name
-                for model in models {
-                    if model.name.eq_ignore_ascii_case(&m3d.filepath) {
-                        model_ids.insert(model.id.to_lowercase());
-                    }
-                }
-            }
-        }
-
-        if model_ids.is_empty() {
-            let result = json!({
-                "status": "error",
-                "filepath": filepath,
-                "footprint": footprint_name,
-                "error": "No 3D model references found for this footprint",
-            });
-            return ToolCallResult::error(serde_json::to_string_pretty(&result).unwrap());
-        }
-
-        // Find matching models
-        let matching_models: Vec<&crate::altium::pcblib::EmbeddedModel> = models
-            .iter()
-            .filter(|m| {
-                let id_lower = m.id.to_lowercase();
-                let id_trimmed = id_lower.trim_matches(|c| c == '{' || c == '}');
-                model_ids.contains(&id_lower)
-                    || model_ids.contains(id_trimmed)
-                    || model_ids
-                        .iter()
-                        .any(|mid| mid.trim_matches(|c| c == '{' || c == '}') == id_trimmed)
-            })
-            .copied()
-            .collect();
-
-        if matching_models.is_empty() {
-            let result = json!({
-                "status": "error",
-                "filepath": filepath,
-                "footprint": footprint_name,
-                "error": "Referenced model IDs not found in embedded models",
-                "referenced_ids": model_ids.iter().collect::<Vec<_>>(),
-                "available_model_ids": models.iter().map(|m| &m.id).collect::<Vec<_>>(),
-            });
-            return ToolCallResult::error(serde_json::to_string_pretty(&result).unwrap());
-        }
-
-        // Extract or return the models
-        if matching_models.len() == 1 {
-            // Single model - use standard extraction
-            Self::extract_model_output(filepath, output_path, matching_models[0])
-        } else if let Some(out_dir) = output_path {
-            // Multiple models - extract to directory
-            Self::extract_all_step_models(filepath, Some(out_dir), &matching_models)
-        } else {
-            // Multiple models, no output - return info
-            let model_info: Vec<Value> = matching_models
-                .iter()
-                .map(|m| {
-                    json!({
-                        "id": m.id,
-                        "name": m.name,
-                        "size_bytes": m.data.len(),
-                    })
-                })
-                .collect();
-
-            let result = json!({
-                "status": "list",
-                "filepath": filepath,
-                "footprint": footprint_name,
-                "message": "Multiple models found for footprint. Specify 'output_path' to extract all.",
-                "model_count": matching_models.len(),
-                "models": model_info,
-            });
-            ToolCallResult::text(serde_json::to_string_pretty(&result).unwrap())
-        }
-    }
-
-    /// Helper to output extracted model data (to file or base64).
-    fn extract_model_output(
-        filepath: &str,
-        output_path: Option<&str>,
-        model: &crate::altium::pcblib::EmbeddedModel,
-    ) -> ToolCallResult {
-        output_path.map_or_else(
-            || {
-                // Return as base64
-                let base64_data = BASE64_STANDARD.encode(&model.data);
-                let result = json!({
-                    "status": "success",
-                    "filepath": filepath,
-                    "model_id": model.id,
-                    "model_name": model.name,
-                    "size_bytes": model.data.len(),
-                    "encoding": "base64",
-                    "data": base64_data,
-                });
-                ToolCallResult::text(serde_json::to_string_pretty(&result).unwrap())
-            },
-            |out_path| {
-                // Write to file
-                match std::fs::write(out_path, &model.data) {
-                    Ok(()) => {
-                        let result = json!({
-                            "status": "success",
-                            "filepath": filepath,
-                            "output_path": out_path,
-                            "model_id": model.id,
-                            "model_name": model.name,
-                            "size_bytes": model.data.len(),
-                            "message": format!("STEP model extracted to '{}'", out_path),
-                        });
-                        ToolCallResult::text(serde_json::to_string_pretty(&result).unwrap())
-                    }
-                    Err(e) => {
-                        let result = json!({
-                            "status": "error",
-                            "filepath": filepath,
-                            "output_path": out_path,
-                            "error": format!("Failed to write file: {}", e),
-                        });
-                        ToolCallResult::error(serde_json::to_string_pretty(&result).unwrap())
-                    }
-                }
-            },
-        )
-    }
-
-    // ==================== Library Diff Tools ====================
-
-    /// Compares two Altium library files and reports differences.
-    fn call_diff_libraries(&self, arguments: &Value) -> ToolCallResult {
-        let Some(filepath_a) = arguments.get("filepath_a").and_then(Value::as_str) else {
-            return ToolCallResult::error("Missing required parameter: filepath_a");
-        };
-
-        let Some(filepath_b) = arguments.get("filepath_b").and_then(Value::as_str) else {
-            return ToolCallResult::error("Missing required parameter: filepath_b");
-        };
-
-        // Validate both paths
-        if let Err(e) = self.validate_path(filepath_a) {
-            return ToolCallResult::error(e);
-        }
-        if let Err(e) = self.validate_path(filepath_b) {
-            return ToolCallResult::error(e);
-        }
-
-        // Determine file types from extensions
-        let path_a = std::path::Path::new(filepath_a);
-        let path_b = std::path::Path::new(filepath_b);
-
-        let ext_a = path_a
-            .extension()
-            .and_then(|e| e.to_str())
-            .map(str::to_lowercase);
-        let ext_b = path_b
-            .extension()
-            .and_then(|e| e.to_str())
-            .map(str::to_lowercase);
-
-        // Ensure both files are the same type
-        if ext_a != ext_b {
-            let result = json!({
-                "status": "error",
-                "error": format!("File types must match. Got '{}' and '{}'.",
-                    ext_a.as_deref().unwrap_or("unknown"),
-                    ext_b.as_deref().unwrap_or("unknown"))
-            });
-            return ToolCallResult::error(serde_json::to_string_pretty(&result).unwrap());
-        }
-
-        match ext_a.as_deref() {
-            Some("pcblib") => Self::diff_pcblibs(filepath_a, filepath_b),
-            Some("schlib") => Self::diff_schlibs(filepath_a, filepath_b),
-            _ => {
-                let result = json!({
-                    "status": "error",
-                    "error": "Unknown file type. Expected .PcbLib or .SchLib extension.",
-                });
-                ToolCallResult::error(serde_json::to_string_pretty(&result).unwrap())
-            }
-        }
-    }
-
-    /// Compares two `PcbLib` files.
-    #[allow(clippy::too_many_lines)]
-    fn diff_pcblibs(filepath_a: &str, filepath_b: &str) -> ToolCallResult {
-        use crate::altium::PcbLib;
-        use std::collections::HashSet;
-
-        // Read both libraries
-        let lib_a = match PcbLib::open(filepath_a) {
-            Ok(lib) => lib,
-            Err(e) => {
-                let result = json!({
-                    "status": "error",
-                    "error": format!("Failed to read '{}': {e}", crate::altium::error::sanitise_path_for_client(std::path::Path::new(filepath_a))),
-                });
-                return ToolCallResult::error(serde_json::to_string_pretty(&result).unwrap());
-            }
-        };
-
-        let lib_b = match PcbLib::open(filepath_b) {
-            Ok(lib) => lib,
-            Err(e) => {
-                let result = json!({
-                    "status": "error",
-                    "error": format!("Failed to read '{}': {e}", crate::altium::error::sanitise_path_for_client(std::path::Path::new(filepath_b))),
-                });
-                return ToolCallResult::error(serde_json::to_string_pretty(&result).unwrap());
-            }
-        };
-
-        // Get component names from both libraries
-        let names_a: HashSet<String> = lib_a.iter().map(|f| f.name.clone()).collect();
-        let names_b: HashSet<String> = lib_b.iter().map(|f| f.name.clone()).collect();
-
-        // Find added, removed, and common components
-        let added: Vec<&str> = names_b.difference(&names_a).map(String::as_str).collect();
-        let removed: Vec<&str> = names_a.difference(&names_b).map(String::as_str).collect();
-        let common: Vec<&str> = names_a.intersection(&names_b).map(String::as_str).collect();
-
-        // Check for modifications in common components
-        let mut modified: Vec<Value> = Vec::new();
-
-        for name in &common {
-            let fp_a = lib_a.get(name).unwrap();
-            let fp_b = lib_b.get(name).unwrap();
-
-            let mut changes: Vec<String> = Vec::new();
-
-            // Compare descriptions
-            if fp_a.description != fp_b.description {
-                changes.push(format!(
-                    "description: '{}' -> '{}'",
-                    fp_a.description, fp_b.description
-                ));
-            }
-
-            // Compare primitive counts
-            if fp_a.pads.len() != fp_b.pads.len() {
-                changes.push(format!(
-                    "pad_count: {} -> {}",
-                    fp_a.pads.len(),
-                    fp_b.pads.len()
-                ));
-            }
-            if fp_a.tracks.len() != fp_b.tracks.len() {
-                changes.push(format!(
-                    "track_count: {} -> {}",
-                    fp_a.tracks.len(),
-                    fp_b.tracks.len()
-                ));
-            }
-            if fp_a.arcs.len() != fp_b.arcs.len() {
-                changes.push(format!(
-                    "arc_count: {} -> {}",
-                    fp_a.arcs.len(),
-                    fp_b.arcs.len()
-                ));
-            }
-            if fp_a.regions.len() != fp_b.regions.len() {
-                changes.push(format!(
-                    "region_count: {} -> {}",
-                    fp_a.regions.len(),
-                    fp_b.regions.len()
-                ));
-            }
-            if fp_a.text.len() != fp_b.text.len() {
-                changes.push(format!(
-                    "text_count: {} -> {}",
-                    fp_a.text.len(),
-                    fp_b.text.len()
-                ));
-            }
-
-            // Compare 3D model presence (external references)
-            let has_model_a = fp_a.model_3d.is_some();
-            let has_model_b = fp_b.model_3d.is_some();
-            if has_model_a != has_model_b {
-                changes.push(format!(
-                    "external_3d_model: {} -> {}",
-                    if has_model_a { "yes" } else { "no" },
-                    if has_model_b { "yes" } else { "no" }
-                ));
-            }
-
-            // Compare embedded 3D bodies
-            if fp_a.component_bodies.len() != fp_b.component_bodies.len() {
-                changes.push(format!(
-                    "component_body_count: {} -> {}",
-                    fp_a.component_bodies.len(),
-                    fp_b.component_bodies.len()
-                ));
-            }
-
-            if !changes.is_empty() {
-                modified.push(json!({
-                    "name": name,
-                    "changes": changes,
-                }));
-            }
-        }
-
-        let result = json!({
-            "status": "success",
-            "file_type": "PcbLib",
-            "filepath_a": filepath_a,
-            "filepath_b": filepath_b,
-            "summary": {
-                "components_in_a": lib_a.len(),
-                "components_in_b": lib_b.len(),
-                "added_count": added.len(),
-                "removed_count": removed.len(),
-                "modified_count": modified.len(),
-                "unchanged_count": common.len() - modified.len(),
-            },
-            "added": added,
-            "removed": removed,
-            "modified": modified,
-        });
-
-        ToolCallResult::text(serde_json::to_string_pretty(&result).unwrap())
-    }
-
-    /// Compares two `SchLib` files.
-    #[allow(clippy::too_many_lines)]
-    fn diff_schlibs(filepath_a: &str, filepath_b: &str) -> ToolCallResult {
-        use crate::altium::SchLib;
-        use std::collections::HashSet;
-
-        // Read both libraries
-        let lib_a = match SchLib::open(filepath_a) {
-            Ok(lib) => lib,
-            Err(e) => {
-                let result = json!({
-                    "status": "error",
-                    "error": format!("Failed to read '{}': {e}", crate::altium::error::sanitise_path_for_client(std::path::Path::new(filepath_a))),
-                });
-                return ToolCallResult::error(serde_json::to_string_pretty(&result).unwrap());
-            }
-        };
-
-        let lib_b = match SchLib::open(filepath_b) {
-            Ok(lib) => lib,
-            Err(e) => {
-                let result = json!({
-                    "status": "error",
-                    "error": format!("Failed to read '{}': {e}", crate::altium::error::sanitise_path_for_client(std::path::Path::new(filepath_b))),
-                });
-                return ToolCallResult::error(serde_json::to_string_pretty(&result).unwrap());
-            }
-        };
-
-        // Get component names from both libraries
-        let names_a: HashSet<String> = lib_a.iter().map(|s| s.name.clone()).collect();
-        let names_b: HashSet<String> = lib_b.iter().map(|s| s.name.clone()).collect();
-
-        // Find added, removed, and common components
-        let added: Vec<&str> = names_b.difference(&names_a).map(String::as_str).collect();
-        let removed: Vec<&str> = names_a.difference(&names_b).map(String::as_str).collect();
-        let common: Vec<&str> = names_a.intersection(&names_b).map(String::as_str).collect();
-
-        // Check for modifications in common components
-        let mut modified: Vec<Value> = Vec::new();
-
-        for name in &common {
-            let sym_a = lib_a.get(name).unwrap();
-            let sym_b = lib_b.get(name).unwrap();
-
-            let mut changes: Vec<String> = Vec::new();
-
-            // Compare descriptions
-            if sym_a.description != sym_b.description {
-                changes.push(format!(
-                    "description: '{}' -> '{}'",
-                    sym_a.description, sym_b.description
-                ));
-            }
-
-            // Compare designators
-            if sym_a.designator != sym_b.designator {
-                changes.push(format!(
-                    "designator: '{}' -> '{}'",
-                    sym_a.designator, sym_b.designator
-                ));
-            }
-
-            // Compare primitive counts
-            if sym_a.pins.len() != sym_b.pins.len() {
-                changes.push(format!(
-                    "pin_count: {} -> {}",
-                    sym_a.pins.len(),
-                    sym_b.pins.len()
-                ));
-            }
-            if sym_a.rectangles.len() != sym_b.rectangles.len() {
-                changes.push(format!(
-                    "rectangle_count: {} -> {}",
-                    sym_a.rectangles.len(),
-                    sym_b.rectangles.len()
-                ));
-            }
-            if sym_a.lines.len() != sym_b.lines.len() {
-                changes.push(format!(
-                    "line_count: {} -> {}",
-                    sym_a.lines.len(),
-                    sym_b.lines.len()
-                ));
-            }
-            if sym_a.polylines.len() != sym_b.polylines.len() {
-                changes.push(format!(
-                    "polyline_count: {} -> {}",
-                    sym_a.polylines.len(),
-                    sym_b.polylines.len()
-                ));
-            }
-            if sym_a.arcs.len() != sym_b.arcs.len() {
-                changes.push(format!(
-                    "arc_count: {} -> {}",
-                    sym_a.arcs.len(),
-                    sym_b.arcs.len()
-                ));
-            }
-            if sym_a.footprints.len() != sym_b.footprints.len() {
-                changes.push(format!(
-                    "footprint_count: {} -> {}",
-                    sym_a.footprints.len(),
-                    sym_b.footprints.len()
-                ));
-            }
-
-            if !changes.is_empty() {
-                modified.push(json!({
-                    "name": name,
-                    "changes": changes,
-                }));
-            }
-        }
-
-        let result = json!({
-            "status": "success",
-            "file_type": "SchLib",
-            "filepath_a": filepath_a,
-            "filepath_b": filepath_b,
-            "summary": {
-                "components_in_a": lib_a.len(),
-                "components_in_b": lib_b.len(),
-                "added_count": added.len(),
-                "removed_count": removed.len(),
-                "modified_count": modified.len(),
-                "unchanged_count": common.len() - modified.len(),
-            },
-            "added": added,
-            "removed": removed,
-            "modified": modified,
-        });
 
         ToolCallResult::text(serde_json::to_string_pretty(&result).unwrap())
     }
