+++ src/mcp/tools/diff.rs
@@ -1,0 +1,341 @@
+//! Library diff tools, split from `server.rs`.
+
+use serde_json::{json, Value};
+
+use crate::mcp::server::{McpServer, ToolCallResult};
+
+impl McpServer {
+    // ==================== Library Diff Tools ====================
+
+    /// Compares two Altium library files and reports differences.
+    pub(crate) fn call_diff_libraries(&self, arguments: &Value) -> ToolCallResult {
+        let Some(filepath_a) = arguments.get("filepath_a").and_then(Value::as_str) else {
+            return ToolCallResult::error("Missing required parameter: filepath_a");
+        };
+
+        let Some(filepath_b) = arguments.get("filepath_b").and_then(Value::as_str) else {
+            return ToolCallResult::error("Missing required parameter: filepath_b");
+        };
+
+        // Validate both paths
+        if let Err(e) = self.validate_path(filepath_a) {
+            return ToolCallResult::error(e);
+        }
+        if let Err(e) = self.validate_path(filepath_b) {
+            return ToolCallResult::error(e);
+        }
+
+        // Determine file types from extensions
+        let path_a = std::path::Path::new(filepath_a);
+        let path_b = std::path::Path::new(filepath_b);
+
+        let ext_a = path_a
+            .extension()
+            .and_then(|e| e.to_str())
+            .map(str::to_lowercase);
+        let ext_b = path_b
+            .extension()
+            .and_then(|e| e.to_str())
+            .map(str::to_lowercase);
+
+        // Ensure both files are the same type
+        if ext_a != ext_b {
+            let result = json!({
+                "status": "error",
+                "error": format!("File types must match. Got '{}' and '{}'.",
+                    ext_a.as_deref().unwrap_or("unknown"),
+                    ext_b.as_deref().unwrap_or("unknown"))
+            });
+            return ToolCallResult::error(serde_json::to_string_pretty(&result).unwrap());
+        }
+
+        match ext_a.as_deref() {
+            Some("pcblib") => Self::diff_pcblibs(filepath_a, filepath_b),
+            Some("schlib") => Self::diff_schlibs(filepath_a, filepath_b),
+            _ => {
+                let result = json!({
+                    "status": "error",
+                    "error": "Unknown file type. Expected .PcbLib or .SchLib extension.",
+                });
+                ToolCallResult::error(serde_json::to_string_pretty(&result).unwrap())
+            }
+        }
+    }
+
+    /// Compares two `PcbLib` files.
+    #[allow(clippy::too_many_lines)]
+    pub(crate) fn diff_pcblibs(filepath_a: &str, filepath_b: &str) -> ToolCallResult {
+        use crate::altium::PcbLib;
+        use std::collections::HashSet;
+
+        // Read both libraries
+        let lib_a = match PcbLib::open(filepath_a) {
+            Ok(lib) => lib,
+            Err(e) => {
+                let result = json!({
+                    "status": "error",
+                    "error": format!("Failed to read '{}': {e}", crate::altium::error::sanitise_path_for_client(std::path::Path::new(filepath_a))),
+                });
+                return ToolCallResult::error(serde_json::to_string_pretty(&result).unwrap());
+            }
+        };
+
+        let lib_b = match PcbLib::open(filepath_b) {
+            Ok(lib) => lib,
+            Err(e) => {
+                let result = json!({
+                    "status": "error",
+                    "error": format!("Failed to read '{}': {e}", crate::altium::error::sanitise_path_for_client(std::path::Path::new(filepath_b))),
+                });
+                return ToolCallResult::error(serde_json::to_string_pretty(&result).unwrap());
+            }
+        };
+
+        // Get component names from both libraries
+        let names_a: HashSet<String> = lib_a.iter().map(|f| f.name.clone()).collect();
+        let names_b: HashSet<String> = lib_b.iter().map(|f| f.name.clone()).collect();
+
+        // Find added, removed, and common components
+        let added: Vec<&str> = names_b.difference(&names_a).map(String::as_str).collect();
+        let removed: Vec<&str> = names_a.difference(&names_b).map(String::as_str).collect();
+        let common: Vec<&str> = names_a.intersection(&names_b).map(String::as_str).collect();
+
+        // Check for modifications in common components
+        let mut modified: Vec<Value> = Vec::new();
+
+        for name in &common {
+            let fp_a = lib_a.get(name).unwrap();
+            let fp_b = lib_b.get(name).unwrap();
+
+            let mut changes: Vec<String> = Vec::new();
+
+            // Compare descriptions
+            if fp_a.description != fp_b.description {
+                changes.push(format!(
+                    "description: '{}' -> '{}'",
+                    fp_a.description, fp_b.description
+                ));
+            }
+
+            // Compare primitive counts
+            if fp_a.pads.len() != fp_b.pads.len() {
+                changes.push(format!(
+                    "pad_count: {} -> {}",
+                    fp_a.pads.len(),
+                    fp_b.pads.len()
+                ));
+            }
+            if fp_a.tracks.len() != fp_b.tracks.len() {
+                changes.push(format!(
+                    "track_count: {} -> {}",
+                    fp_a.tracks.len(),
+                    fp_b.tracks.len()
+                ));
+            }
+            if fp_a.arcs.len() != fp_b.arcs.len() {
+                changes.push(format!(
+                    "arc_count: {} -> {}",
+                    fp_a.arcs.len(),
+                    fp_b.arcs.len()
+                ));
+            }
+            if fp_a.regions.len() != fp_b.regions.len() {
+                changes.push(format!(
+                    "region_count: {} -> {}",
+                    fp_a.regions.len(),
+                    fp_b.regions.len()
+                ));
+            }
+            if fp_a.text.len() != fp_b.text.len() {
+                changes.push(format!(
+                    "text_count: {} -> {}",
+                    fp_a.text.len(),
+                    fp_b.text.len()
+                ));
+            }
+
+            // Compare 3D model presence (external references)
+            let has_model_a = fp_a.model_3d.is_some();
+            let has_model_b = fp_b.model_3d.is_some();
+            if has_model_a != has_model_b {
+                changes.push(format!(
+                    "external_3d_model: {} -> {}",
+                    if has_model_a { "yes" } else { "no" },
+                    if has_model_b { "yes" } else { "no" }
+                ));
+            }
+
+            // Compare embedded 3D bodies
+            if fp_a.component_bodies.len() != fp_b.component_bodies.len() {
+                changes.push(format!(
+                    "component_body_count: {} -> {}",
+                    fp_a.component_bodies.len(),
+                    fp_b.component_bodies.len()
+                ));
+            }
+
+            if !changes.is_empty() {
+                modified.push(json!({
+                    "name": name,
+                    "changes": changes,
+                }));
+            }
+        }
+
+        let result = json!({
+            "status": "success",
+            "file_type": "PcbLib",
+            "filepath_a": filepath_a,
+            "filepath_b": filepath_b,
+            "summary": {
+                "components_in_a": lib_a.len(),
+                "components_in_b": lib_b.len(),
+                "added_count": added.len(),
+                "removed_count": removed.len(),
+                "modified_count": modified.len(),
+                "unchanged_count": common.len() - modified.len(),
+            },
+            "added": added,
+            "removed": removed,
+            "modified": modified,
+        });
+
+        ToolCallResult::text(serde_json::to_string_pretty(&result).unwrap())
+    }
+
+    /// Compares two `SchLib` files.
+    #[allow(clippy::too_many_lines)]
+    pub(crate) fn diff_schlibs(filepath_a: &str, filepath_b: &str) -> ToolCallResult {
+        use crate::altium::SchLib;
+        use std::collections::HashSet;
+
+        // Read both libraries
+        let lib_a = match SchLib::open(filepath_a) {
+            Ok(lib) => lib,
+            Err(e) => {
+                let result = json!({
+                    "status": "error",
+                    "error": format!("Failed to read '{}': {e}", crate::altium::error::sanitise_path_for_client(std::path::Path::new(filepath_a))),
+                });
+                return ToolCallResult::error(serde_json::to_string_pretty(&result).unwrap());
+            }
+        };
+
+        let lib_b = match SchLib::open(filepath_b) {
+            Ok(lib) => lib,
+            Err(e) => {
+                let result = json!({
+                    "status": "error",
+                    "error": format!("Failed to read '{}': {e}", crate::altium::error::sanitise_path_for_client(std::path::Path::new(filepath_b))),
+                });
+                return ToolCallResult::error(serde_json::to_string_pretty(&result).unwrap());
+            }
+        };
+
+        // Get component names from both libraries
+        let names_a: HashSet<String> = lib_a.iter().map(|s| s.name.clone()).collect();
+        let names_b: HashSet<String> = lib_b.iter().map(|s| s.name.clone()).collect();
+
+        // Find added, removed, and common components
+        let added: Vec<&str> = names_b.difference(&names_a).map(String::as_str).collect();
+        let removed: Vec<&str> = names_a.difference(&names_b).map(String::as_str).collect();
+        let common: Vec<&str> = names_a.intersection(&names_b).map(String::as_str).collect();
+
+        // Check for modifications in common components
+        let mut modified: Vec<Value> = Vec::new();
+
+        for name in &common {
+            let sym_a = lib_a.get(name).unwrap();
+            let sym_b = lib_b.get(name).unwrap();
+
+            let mut changes: Vec<String> = Vec::new();
+
+            // Compare descriptions
+            if sym_a.description != sym_b.description {
+                changes.push(format!(
+                    "description: '{}' -> '{}'",
+                    sym_a.description, sym_b.description
+                ));
+            }
+
+            // Compare designators
+            if sym_a.designator != sym_b.designator {
+                changes.push(format!(
+                    "designator: '{}' -> '{}'",
+                    sym_a.designator, sym_b.designator
+                ));
+            }
+
+            // Compare primitive counts
+            if sym_a.pins.len() != sym_b.pins.len() {
+                changes.push(format!(
+                    "pin_count: {} -> {}",
+                    sym_a.pins.len(),
+                    sym_b.pins.len()
+                ));
+            }
+            if sym_a.rectangles.len() != sym_b.rectangles.len() {
+                changes.push(format!(
+                    "rectangle_count: {} -> {}",
+                    sym_a.rectangles.len(),
+                    sym_b.rectangles.len()
+                ));
+            }
+            if sym_a.lines.len() != sym_b.lines.len() {
+                changes.push(format!(
+                    "line_count: {} -> {}",
+                    sym_a.lines.len(),
+                    sym_b.lines.len()
+                ));
+            }
+            if sym_a.polylines.len() != sym_b.polylines.len() {
+                changes.push(format!(
+                    "polyline_count: {} -> {}",
+                    sym_a.polylines.len(),
+                    sym_b.polylines.len()
+                ));
+            }
+            if sym_a.arcs.len() != sym_b.arcs.len() {
+                changes.push(format!(
+                    "arc_count: {} -> {}",
+                    sym_a.arcs.len(),
+                    sym_b.arcs.len()
+                ));
+            }
+            if sym_a.footprints.len() != sym_b.footprints.len() {
+                changes.push(format!(
+                    "footprint_count: {} -> {}",
+                    sym_a.footprints.len(),
+                    sym_b.footprints.len()
+                ));
+            }
+
+            if !changes.is_empty() {
+                modified.push(json!({
+                    "name": name,
+                    "changes": changes,
+                }));
+            }
+        }
+
+        let result = json!({
+            "status": "success",
+            "file_type": "SchLib",
+            "filepath_a": filepath_a,
+            "filepath_b": filepath_b,
+            "summary": {
+                "components_in_a": lib_a.len(),
+                "components_in_b": lib_b.len(),
+                "added_count": added.len(),
+                "removed_count": removed.len(),
+                "modified_count": modified.len(),
+                "unchanged_count": common.len() - modified.len(),
+            },
+            "added": added,
+            "removed": removed,
+            "modified": modified,
+        });
+
+        ToolCallResult::text(serde_json::to_string_pretty(&result).unwrap())
+    }
+}
